Make yomichan detect separate captions as separate sentences
When yomichan scans a word and adds it to anki, it tries to get the
surrounding content to populate the sentence field it adds to anki.
But whenever I've tried this with the sidebar on videobook, it
always grabs like 10 surrounding captions since it thinks they're all
the same sentence. This change makes yomichan only pull in the content
for the caption you found the word in.