Only disable quotes when the user selects multiple captions
Trying to make it so that if a single caption has newlines, yomichan
will still grab just the caption's sentence when you extract to anki.
The only thing we really need to be careful of is to make sure quotes
don't get accidentally copied to the clipboard when a user highlights
text from multiple captions.